# openapi.yaml
openapi: 3.0.3
info:
  title: Nautix API
  version: 0.1.0
servers:
  - url: http://localhost:8000
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200': { description: OK }
  /search:
    get:
      summary: Search ferry schedules
      parameters:
        - in: query
          name: origin_port_id
          required: true
          schema: { type: string }
        - in: query
          name: dest_port_id
          required: true
          schema: { type: string }
        - in: query
          name: date
          required: true
          schema: { type: string, format: date }
        - in: query
          name: pax
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200': { description: OK }
  /bookings:
    post:
      summary: Create booking (intent -> confirmed after payment)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [schedule_id, passengers]
              properties:
                schedule_id: { type: string }
                passengers:
                  type: array
                  items: { type: object, properties: { name: { type: string } } }
                vehicle:
                  type: object
                  properties: { type: { type: string } }
                addons: { type: array, items: { type: string } }
      responses:
        '201': { description: Created }
  /tickets/{ticketId}:
    get:
      summary: Get ticket detail (QR token included)
      parameters:
        - in: path
          name: ticketId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /scan:
    post:
      summary: Validate ticket scan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                qr_token: { type: string }
      responses:
        '200': { description: OK } 